major: upgrade to HotShot 0.5.10, reorg contracts/rust to avoid cyclic dep
