--- contracts/rust/Cargo.toml
@@ -1,51 +1,0 @@
-[package]
-name = "hotshot_contract"
-version = "0.1.0"
-authors = ["Espresso Systems <hello@espressosys.com>"]
-edition = "2021"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.79"
-ark-bn254 = "0.4"
-ark-ec = "0.4"
-ark-ed-on-bn254 = "0.4.0"
-ark-ff = "0.4"
-ark-poly = "0.4"
-ark-serialize = "0.4"
-ark-std = { version = "0.4", default-features = false }
-async-std = "1.12.0"
-async-trait = "0.1.74"
-clap = { version = "^4.4", features = ["derive"] }
-contract-bindings = { path = "../../contract-bindings" }
-crs = { git = "https://github.com/alxiong/crs" }
-diff-test-bn254 = { git = "https://github.com/EspressoSystems/solidity-bn254.git" }
-digest = { version = "0.10", default-features = false, features = ["alloc"] }
-ethereum-types = { version = "^0.14", features = ["impl-serde"] }
-ethers = { version = "2.0.4" }
-ethers-providers = "2.0.4"
-hex = "0.4.3"
-hotshot-stake-table = { workspace = true }
-hotshot-state-prover = { workspace = true }
-hotshot-types = { workspace = true }
-itertools = "0.12.0"
-jf-plonk = { workspace = true }
-jf-primitives = { workspace = true }
-jf-relation = { workspace = true }
-jf-utils = { workspace = true }
-num-bigint = { version = "0.4", default-features = false }
-num-traits = { version = "0.2", default-features = false }
-sequencer-utils = { path = "../../utils" }
-sha3 = { version = "0.10.8", default-features = false }
-
-[[bin]]
-name = "diff-test"
-path = "src/bin/diff_test.rs"
-
-[[bin]]
-name = "gen-vk-contract"
-path = "src/bin/gen_vk_contract.rs"
-
-[[bin]]
-name = "eval-domain"
-path = "src/bin/eval_domain.rs"
