+++ contracts/rust/adapter/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "hotshot-contract-adapter"
+description = "Cross-domain (between Solidity and Rust) utilities for type conversion and testing"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+ark-bn254 = { workspace = true }
+ark-ec = { workspace = true }
+ark-ed-on-bn254 = { workspace = true }
+ark-ff = { workspace = true }
+ark-poly = { workspace = true }
+ark-serialize = { workspace = true }
+ark-std = { workspace = true }
+contract-bindings = { path = "../../../contract-bindings" }
+crs = { git = "https://github.com/alxiong/crs" }
+diff-test-bn254 = { git = "https://github.com/EspressoSystems/solidity-bn254.git" }
+ethers = { version = "2.0.4" }
+hotshot-types = { workspace = true }
+jf-plonk = { workspace = true }
+jf-primitives = { workspace = true }
+jf-utils = { workspace = true }
+num-bigint = { version = "0.4", default-features = false }
+num-traits = { version = "0.2", default-features = false }
+
+[dev-dependencies]
+hotshot-stake-table = { workspace = true }
+
+[[bin]]
+name = "eval-domain"
+path = "src/bin/eval_domain.rs"
