+++ contracts/rust/adapter/src/bin/eval_domain.rs
