+++ contracts/rust/adapter/src/lib.rs
@@ -1,0 +1,7 @@
+//! Cross-domain (between Solidity and Rust) utilities for type conversion and testing
+
+pub mod jellyfish;
+pub mod light_client;
+
+// Archived, legacy helpers and tests, to be removed soon. not included, reference/read only
+// mod archived
