+++ contracts/rust/diff-test/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "diff-test-hotshot"
+description = "Helpers and mocks for Forge-powered differential tests on HotShot-related contracts"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+ark-bn254 = { workspace = true }
+ark-ec = { workspace = true }
+ark-ed-on-bn254 = { workspace = true }
+ark-ff = { workspace = true }
+ark-poly = { workspace = true }
+ark-std = { workspace = true }
+clap = { version = "^4.4", features = ["derive"] }
+crs = { git = "https://github.com/alxiong/crs" }
+diff-test-bn254 = { git = "https://github.com/EspressoSystems/solidity-bn254.git" }
+digest = { version = "0.10", default-features = false, features = ["alloc"] }
+ethers = { version = "2.0.4" }
+hotshot-contract-adapter = { path = "../adapter" }
+hotshot-stake-table = { workspace = true }
+hotshot-state-prover = { path = "../../../hotshot-state-prover" }
+hotshot-types = { workspace = true }
+itertools = { workspace = true }
+jf-plonk = { workspace = true }
+jf-primitives = { workspace = true }
+jf-relation = { workspace = true }
+jf-utils = { workspace = true }
+sha3 = { version = "0.10.8", default-features = false }
+
+[[bin]]
+name = "diff-test"
+path = "src/main.rs"
