--- contracts/rust/src/lib.rs
@@ -1,4 +1,0 @@
-mod helpers;
-mod hotshot_contract;
-pub mod jf_helpers;
-pub mod light_client;
