+++ hotshot-state-prover/Cargo.toml
@@ -19,9 +19,11 @@ async-trait = { workspace = true }
 blake3 = "1.5"
 clap = { workspace = true }
 cld = { workspace = true }
+contract-bindings = { path = "../contract-bindings" }
 crs = { git = "https://github.com/alxiong/crs" }
 displaydoc = { version = "0.2.3", default-features = false }
 ethers = { workspace = true }
+hotshot-contract-adapter = { path = "../contracts/rust/adapter" }
 hotshot-stake-table = { workspace = true }
 hotshot-types = { workspace = true }
 jf-plonk = { workspace = true }
