+++ hotshot-state-prover/src/bin/state-prover.rs
@@ -1,7 +1,12 @@
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
 use cld::ClDuration;
-use hotshot_state_prover::service::{key_gen, run_prover_once, run_prover_service};
+use ethers::providers::{Http, Middleware, Provider};
+use ethers::signers::{coins_bip39::English, MnemonicBuilder, Signer};
+use ethers::types::Address;
+use hotshot_state_prover::service::{
+    key_gen, run_prover_once, run_prover_service, StateProverConfig,
+};
 use snafu::Snafu;
 use std::{path::PathBuf, str::FromStr as _, time::Duration};
 use url::Url;
@@ -16,9 +21,14 @@ struct Args {
     #[clap(long = "key-gen", action)]
     keygen: bool,
 
-    /// The frequency of updating the light client state
-    #[clap(short, long = "freq", value_parser = parse_duration, default_value = "10m", env = "ESPRESSO_STATE_PROVER_FREQUENCY")]
-    frequency: Duration,
+    /// Path to the proving key
+    #[clap(
+        short = 'k',
+        long = "key",
+        default_value = "key",
+        env = "ESPRESSOS_STATE_PROVING_KEY"
+    )]
+    proving_key_path: PathBuf,
 
     /// Url of the state relay server
     #[clap(
@@ -28,14 +38,29 @@ struct Args {
     )]
     relay_server: Url,
 
-    /// Path to the proving key
+    /// The frequency of updating the light client state, expressed in update interval
+    #[clap(short, long = "freq", value_parser = parse_duration, default_value = "10m", env = "ESPRESSO_STATE_PROVER_UPDATE_INTERVAL")]
+    update_interval: Duration,
+
+    /// URL of layer 1 Ethereum JSON-RPC provider.
+    #[clap(long, env = "ESPRESSO_SEQUENCER_L1_PROVIDER")]
+    pub l1_provider: Url,
+
+    /// Address of LightClient contract on layer 1.
+    #[clap(long, env = "ESPRESSO_SEQUENCER_LIGHTCLIENT_ADDRESS")]
+    pub light_client_address: Address,
+
+    /// Mnemonic phrase for a funded Ethereum wallet.
+    #[clap(long, env = "ESPRESSO_SEQUENCER_ETH_MNEMONIC", default_value = None)]
+    pub eth_mnemonic: String,
+
+    /// Index of a funded account derived from eth-mnemonic.
     #[clap(
-        short = 'k',
-        long = "key",
-        default_value = "key",
-        env = "ESPRESSOS_STATE_PROVING_KEY"
+        long,
+        env = "ESPRESSO_SEQUENCER_ETH_ACCOUNT_INDEX",
+        default_value = "0"
     )]
-    proving_key_path: PathBuf,
+    pub eth_account_index: u32,
 }
 
 #[derive(Clone, Debug, Snafu)]
@@ -58,14 +83,34 @@ async fn main() {
 
     let args = Args::parse();
 
+    // prepare config for state prover from user options
+    let provider = Provider::<Http>::try_from(args.l1_provider.to_string()).unwrap();
+    let chain_id = provider.get_chainid().await.unwrap().as_u64();
+    let config = StateProverConfig {
+        proving_key_path: args.proving_key_path.clone(),
+        relay_server: args.relay_server.clone(),
+        update_interval: args.update_interval,
+        l1_provider: args.l1_provider.clone(),
+        light_client_address: args.light_client_address,
+        eth_signing_key: MnemonicBuilder::<English>::default()
+            .phrase(args.eth_mnemonic.as_str())
+            .index(args.eth_account_index)
+            .expect("error building wallet")
+            .build()
+            .expect("error opening wallet")
+            .with_chain_id(chain_id)
+            .signer()
+            .clone(),
+    };
+
     if args.keygen {
         // Key gen route
         key_gen(args.proving_key_path)
     } else if args.daemon {
         // Launching the prover service daemon
-        run_prover_service(args.proving_key_path, args.relay_server, args.frequency).await;
+        run_prover_service(config).await;
     } else {
         // Run light client state update once
-        run_prover_once(args.proving_key_path, args.relay_server).await;
+        run_prover_once(config).await;
     }
 }
